+++ lib/coinbase.rb
@@ -14,7 +14,7 @@ require_relative 'coinbase/network'
 require_relative 'coinbase/transfer'
 require_relative 'coinbase/user'
 require_relative 'coinbase/wallet'
-require_relative 'coinbase/signer'
+require_relative 'coinbase/server_signer'
 require 'json'
 
 # The Coinbase SDK.
